Add `git cliff` config
When I started maturin, I chose the https://keepachangelog.com style for the project changelog, and would edit the changelog manually. I now want to change this to automated changelog generation and a slightly improved style.

This PR adds a [git cliff](https://git-cliff.org/) config that can generate the changelog based on commits and PR labels. The style is simplified, The version header doesn't have a link anymore, and the PR link is in parentheses behind the description.

It also adds a release guide to the contributing guidelines, with the goal of simplifying the release workflow.

Blocked on https://github.com/orhun/git-cliff/issues/1189